Broke down appDesign page into section components..
